backend/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.11
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9
pytz==2021.3
sqlparse==0.4.2
gunicorn==20.1.0
django-cors-headers==3.11.0
django-filter==21.1
django-environ==0.4.5
ipython==8.3.0
django-allauth==0.51.0
dj-rest-auth==2.2.5
pyyaml==6.0
redis==4.3.3
django-redis==5.2.0
PyJWT==2.4.0
coverage==6.4.4
django-extensions==3.2.0
celery==5.2.7
python-dateutil==2.8.2
django-ckeditor==6.5.1
django-material-admin==1.8.6
boto3==1.24.39
django-storages==1.12.3
Pillow==9.2.0
drf-api-logger==1.1.11
django-3-jet==1.0.8